+++ src/pbfbench/topics/items.py
@@ -5,6 +5,7 @@ from typing import final
 import pbfbench.abc.topic.description as abc_topic_desc
 import pbfbench.abc.topic.visitor as abc_topic_visitor
 import pbfbench.topics.assembly.description as asm_desc
+import pbfbench.topics.plasmidness.description as plasmidness_desc
 import pbfbench.topics.seeds.description as seeds_desc
 
 # REFACTOR reverse the sense: topic container.???
@@ -17,6 +18,7 @@ class Topics(abc_topic_visitor.Topics):
 
     ASSEMBLY = asm_desc.DESCRIPTION.name()
     SEEDS = seeds_desc.DESCRIPTION.name()
+    PLASMIDNESS = plasmidness_desc.DESCRIPTION.name()
 
     def to_description(self) -> abc_topic_desc.Description:
         """Get topic description."""
@@ -25,3 +27,5 @@ class Topics(abc_topic_visitor.Topics):
                 return asm_desc.DESCRIPTION
             case Topics.SEEDS:
                 return seeds_desc.DESCRIPTION
+            case Topics.PLASMIDNESS:
+                return plasmidness_desc.DESCRIPTION