+++ src/pbfbench/topics/plasmidness/description.py
@@ -1,0 +1,5 @@
+"""Plasmidness topic description."""
+
+import pbfbench.abc.topic.description as abc_topic_desc
+
+DESCRIPTION = abc_topic_desc.Description("PLASMIDNESS", "plasmidness")