+++ src/pbfbench/topics/visitor.py
@@ -3,6 +3,7 @@
 import pbfbench.abc.topic.visitor as abc_topic_visitor
 import pbfbench.topics.assembly.visitor as asm_visitor
 import pbfbench.topics.items as topics_items
+import pbfbench.topics.plasmidness.visitor as plasmidness_visitor
 import pbfbench.topics.seeds.visitor as seeds_visitor
 
 
@@ -13,3 +14,5 @@ def tools(topic: topics_items.Topics) -> type[abc_topic_visitor.Tools]:
             return asm_visitor.Tools
         case topics_items.Topics.SEEDS:
             return seeds_visitor.Tools
+        case topics_items.Topics.PLASMIDNESS:
+            return plasmidness_visitor.Tools